Move Networking into Authentication.Keycloak
