+++ visitz/Visitz/Visitz.csproj
@@ -108,7 +108,6 @@
 	  <Folder Include="Views\" />
 	  <Folder Include="ViewModels\" />
 	  <Folder Include="Services\Authentication\" />
-	  <Folder Include="Services\Networking\" />
 	  <Folder Include="Models\" />
 	  <Folder Include="Resources\Splash\" />
 	  <Folder Include="Resources\Localization\" />
